requirements.txt
aiobotocore==2.7.0
aiohttp==3.9.0
aioitertools==0.11.0
aiosignal==1.3.1
asgiref==3.7.2
async-timeout==4.0.3
attrs==23.1.0
boto3==1.29.3
botocore==1.31.64
certifi==2023.11.17
charset-normalizer==3.3.2
contourpy==1.2.0
cycler==0.12.1
Django==4.2.6
djangorestframework==3.14.0
fonttools==4.44.0
frozenlist==1.4.0
fsspec==2023.10.0
idna==3.4
jmespath==1.0.1
kiwisolver==1.4.5
matplotlib==3.8.1
multidict==6.0.4
numpy==1.26.2
packaging==23.2
pandas==2.1.3
Pillow==10.1.0
pyparsing==3.1.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3.post1
requests==2.31.0
s3fs==2023.10.0
s3transfer==0.7.0
six==1.16.0
sqlparse==0.4.4
typing_extensions==4.8.0
tzdata==2023.3
urllib3==2.0.7
windrose==1.9.0
wrapt==1.16.0
yarl==1.9.2
